backend/requirements.txt
fastapi
uvicorn
PyMuPDF
langchain
matplotlib
openai
pinecone-client
sqlalchemy
requests
beautifulsoup4
